+++ R/future_xmap.R
@@ -7,7 +7,7 @@
 #' @param .progress A logical, for whether or not to print a progress bar for
 #'   multiprocess, multisession, and multicore plans
 #' @param .options The `future` specific options to use with the workers.
-#'   This must be the result from a call to [furrr::future_options()]
+#'   This must be the result from a call to [furrr::furrr_options()]
 #'
 #' @return An atomic vector, list, or data frame, depending on the suffix.
 #'   Atomic vectors and lists will be named if the first element of .l is named.
@@ -25,13 +25,12 @@
 #'   [furrr::future_map()], [furrr::future_map2()], and [furrr::future_pmap()]
 #'   for other parallelized mapping functions.
 #'
-#' @include errors.R
 #' @export
 #'
 #' @example examples/future_xmap.R
 
 future_xmap <- function(
-  .l, .f, ..., .progress = FALSE, .options = furrr::future_options()
+  .l, .f, ..., .progress = FALSE, .options = furrr::furrr_options()
 ) {
   require_furrr()
   furrr::future_pmap(
@@ -43,7 +42,7 @@ future_xmap <- function(
 #' @export
 
 future_xmap_chr <- function(
-  .l, .f, ..., .progress = FALSE, .options = furrr::future_options()
+  .l, .f, ..., .progress = FALSE, .options = furrr::furrr_options()
 ) {
   require_furrr()
   furrr::future_pmap_chr(
@@ -55,7 +54,7 @@ future_xmap_chr <- function(
 #' @export
 
 future_xmap_dbl <- function(
-  .l, .f, ..., .progress = FALSE, .options = furrr::future_options()
+  .l, .f, ..., .progress = FALSE, .options = furrr::furrr_options()
 ) {
   require_furrr()
   furrr::future_pmap_dbl(
@@ -67,10 +66,9 @@ future_xmap_dbl <- function(
 #' @export
 
 future_xmap_dfc <- function(
-  .l, .f, ..., .progress = FALSE, .options = furrr::future_options()
+  .l, .f, ..., .progress = FALSE, .options = furrr::furrr_options()
 ) {
   require_furrr()
-  require_package("dplyr")
   furrr::future_pmap_dfc(
     cross_list(.l), .f, ..., .progress = .progress, .options = .options
   )
@@ -80,10 +78,9 @@ future_xmap_dfc <- function(
 #' @export
 
 future_xmap_dfr <- function(
-  .l, .f, ..., .id = NULL, .progress = FALSE, .options = furrr::future_options()
+  .l, .f, ..., .id = NULL, .progress = FALSE, .options = furrr::furrr_options()
 ) {
   require_furrr()
-  require_package("dplyr")
   furrr::future_pmap_dfr(
     cross_list(.l), .f, ..., .id, .progress = .progress, .options = .options
   )
@@ -93,7 +90,7 @@ future_xmap_dfr <- function(
 #' @export
 
 future_xmap_int <- function(
-  .l, .f, ..., .progress = FALSE, .options = furrr::future_options()
+  .l, .f, ..., .progress = FALSE, .options = furrr::furrr_options()
 ) {
   require_furrr()
   furrr::future_pmap_int(
@@ -105,7 +102,7 @@ future_xmap_int <- function(
 #' @export
 
 future_xmap_lgl <- function(
-  .l, .f, ..., .progress = FALSE, .options = furrr::future_options()
+  .l, .f, ..., .progress = FALSE, .options = furrr::furrr_options()
 ) {
   require_furrr()
   furrr::future_pmap_lgl(
@@ -113,3 +110,26 @@ future_xmap_lgl <- function(
   )
 }
 
+#' @rdname future_xmap
+#' @export
+
+future_xmap_raw <- function(
+  .l, .f, ..., .progress = FALSE, .options = furrr::furrr_options()
+) {
+  require_furrr()
+  furrr::future_pmap_raw(
+    cross_list(.l), .f, ..., .progress = .progress, .options = .options
+  )
+}
+
+#' @rdname future_xmap
+#' @export
+
+future_xwalk <- function(
+  .l, .f, ..., .progress = FALSE, .options = furrr::furrr_options()
+) {
+  require_furrr()
+  furrr::future_pwalk(
+    cross_list(.l), .f, ..., .progress = .progress, .options = .options
+  )
+}