+++ R/future_xmap_mat.R
@@ -12,14 +12,13 @@
 #'
 #'   [future_xmap()] for the underlying functions.
 #'
-#' @include errors.R
 #' @export
 #'
 #' @example examples/future_xmap_mat.R
 
 future_xmap_mat <- function(
   .l, .f, ...,
-  .names = TRUE, .progress = FALSE, .options = furrr::future_options()
+  .names = TRUE, .progress = FALSE, .options = furrr::furrr_options()
 ) {
   require_furrr()
   warn_if_not_matrix(.l)
@@ -34,7 +33,7 @@ future_xmap_mat <- function(
 
 future_xmap_arr <- function(
   .l, .f, ...,
-  .names = TRUE, .progress = FALSE, .options = furrr::future_options()
+  .names = TRUE, .progress = FALSE, .options = furrr::furrr_options()
 ) {
   require_furrr()