+++ tests/testthat/test-future_xmap.R
@@ -42,6 +42,17 @@ for (.e in executors) {
     .xmap        <- xmap_chr(test_l, .f)
     .future_xmap <- future_xmap_chr(test_l, .f)
     expect_equal(.xmap, .future_xmap)
+
+    .f           <- function(x, y) {print(paste(x, y))}
+    .xmap        <- utils::capture.output(xwalk(test_l, .f))
+    .future_xmap <- utils::capture.output(future_xwalk(test_l, .f))
+    expect_equal(.xmap, .future_xmap)
+
+    test_l[[1]]  <- as.raw(test_l[[1]])
+    .f           <- rawShift
+    .xmap        <- xmap_raw(test_l, .f)
+    .future_xmap <- future_xmap_raw(test_l, .f)
+    expect_equal(.xmap, .future_xmap)
   })
 
   test_that(test_msg(.e, "equivalence with df xmap()s"), {
@@ -186,8 +197,8 @@ for (.e in executors) {
       mean(c(.x, .y), na.rm = na.rm)
     }
 
-    test_l_na         <- test_l
-    test_l_na[[1]][1] <- NA
+    test_l_na           <- test_l
+    test_l_na[[1]][[1]] <- NA
 
     .xmap        <- xmap(test_l_na, vec_mean, na.rm = TRUE)
     .future_xmap <- future_xmap(test_l_na, vec_mean, na.rm = TRUE)
@@ -242,8 +253,7 @@ for (.e in executors) {
   })
 
   test_that(
-    test_msg(.e, "Globals in .l are only exported to workers that use them"),
-    {
+    test_msg(.e, "Globals in .l are only exported to workers that use them"), {
       skip_if_not_installed("furrr")
       skip_if_not_installed("future")